Write playGame() & amend game instructions
Moved code around as necessary and wrote the logic to play 5 rounds
and crown a winner.

Note: The game continues to prompt the user and progress even when
the user has inputted an invalid hand entry or cancelled the dialog,
when it should ideally invalidate and reset the game. I have no elegant
solution to this at the moment and it may not be worth the time to try
and implement one. I believe my implementation meets the requirements
of the assignment. It will be something to keep in mind when revisiting
the project later down the line.